Struct google_healthcare1::api::ProjectLocationDatasetHl7V2StoreMessageListCall

pub struct ProjectLocationDatasetHl7V2StoreMessageListCall<'a, C> where
    C: 'a,  { /* fields omitted */ }

Lists all the messages in the given HL7v2 store with support for filtering. Note: HL7v2 messages are indexed asynchronously, so there might be a slight delay between the time a message is created and when it can be found through a filter.

A builder for the locations.datasets.hl7V2Stores.messages.list method supported by a project resource. It is not used directly, but through a ProjectMethods instance.

Example

Instantiate a resource method builder

 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.projects().locations_datasets_hl7_v2_stores_messages_list("parent")
             .view("gubergren")
             .page_token("dolor")
             .page_size(-32)
             .order_by("no")
             .filter("amet.")
             .doit().await;

Implementations

impl<'a, C> ProjectLocationDatasetHl7V2StoreMessageListCall<'a, C> where
C: BorrowMut<Client<HttpsConnector<HttpConnector>, Body>>,

pub async fn doit(self) -> Result<(Response<Body>, ListMessagesResponse)>

Perform the operation you have build so far.

pub fn parent(
self,
new_value: &str
) -> ProjectLocationDatasetHl7V2StoreMessageListCall<'a, C>

Name of the HL7v2 store to retrieve messages from.

Sets the parent path property to the given value.

Even though the property as already been set when instantiating this call, we provide this method for API completeness.

pub fn view(
self,
new_value: &str
) -> ProjectLocationDatasetHl7V2StoreMessageListCall<'a, C>

Specifies the parts of the Message to return in the response. When unspecified, equivalent to BASIC. Setting this to anything other than BASIC with a page_size larger than the default can generate a large response, which impacts the performance of this method.

Sets the view query property to the given value.

pub fn page_token(
self,
new_value: &str
) -> ProjectLocationDatasetHl7V2StoreMessageListCall<'a, C>

The next_page_token value returned from the previous List request, if any.

Sets the page token query property to the given value.

pub fn page_size(
self,
new_value: i32
) -> ProjectLocationDatasetHl7V2StoreMessageListCall<'a, C>

Limit on the number of messages to return in a single response. If not specified, 100 is used. May not be larger than 1000.

Sets the page size query property to the given value.

pub fn order_by(
self,
new_value: &str
) -> ProjectLocationDatasetHl7V2StoreMessageListCall<'a, C>

Orders messages returned by the specified order_by clause. Syntax: https://cloud.google.com/apis/design/design_patterns#sorting_order Fields available for ordering are: * send_time

Sets the order by query property to the given value.

pub fn filter(
self,
new_value: &str
) -> ProjectLocationDatasetHl7V2StoreMessageListCall<'a, C>

Restricts messages returned to those matching a filter. The following syntax is available: * A string field value can be written as text inside quotation marks, for example "query text". The only valid relational operation for text fields is equality (=), where text is searched within the field, rather than having the field be equal to the text. For example, "Comment = great" returns messages with great in the comment field. * A number field value can be written as an integer, a decimal, or an exponential. The valid relational operators for number fields are the equality operator (=), along with the less than/greater than operators (<, <=, >, >=). Note that there is no inequality (!=) operator. You can prepend the NOT operator to an expression to negate it. * A date field value must be written in yyyy-mm-dd form. Fields with date and time use the RFC3339 time format. Leading zeros are required for one-digit months and days. The valid relational operators for date fields are the equality operator (=) , along with the less than/greater than operators (<, <=, >, >=). Note that there is no inequality (!=) operator. You can prepend the NOT operator to an expression to negate it. * Multiple field query expressions can be combined in one query by adding AND or OR operators between the expressions. If a boolean operator appears within a quoted string, it is not treated as special, it's just another part of the character string to be matched. You can prepend the NOT operator to an expression to negate it. Fields/functions available for filtering are: * message_type, from the MSH-9.1 field. For example, NOT message_type = "ADT". * send_date or sendDate, the YYYY-MM-DD date the message was sent in the dataset's time_zone, from the MSH-7 segment. For example, send_date < "2017-01-02". * send_time, the timestamp when the message was sent, using the RFC3339 time format for comparisons, from the MSH-7 segment. For example, send_time < "2017-01-02T00:00:00-05:00". * create_time, the timestamp when the message was created in the HL7v2 store. Use the RFC3339 time format for comparisons. For example, create_time < "2017-01-02T00:00:00-05:00". * send_facility, the care center that the message came from, from the MSH-4 segment. For example, send_facility = "ABC". * PatientId(value, type), which matches if the message lists a patient having an ID of the given value and type in the PID-2, PID-3, or PID-4 segments. For example, PatientId("123456", "MRN"). * labels.x, a string value of the label with key x as set using the Message.labels map. For example, labels."priority"="high". The operator :* can be used to assert the existence of a label. For example, labels."priority":*.

Sets the filter query property to the given value.

pub fn delegate(
self,
new_value: &'a mut dyn Delegate
) -> ProjectLocationDatasetHl7V2StoreMessageListCall<'a, C>

The delegate implementation is consulted whenever there is an intermediate result, or if something goes wrong while executing the actual API request.

It should be used to handle progress information, and to implement a certain level of resilience.

Sets the delegate property to the given value.

pub fn param<T>(
self,
name: T,
value: T
) -> ProjectLocationDatasetHl7V2StoreMessageListCall<'a, C> where
T: AsRef<str>,

Set any additional parameter of the query string used in the request. It should be used to set parameters which are not yet available through their own setters.

Please note that this method must not be used to set any of the known parameters which have their own setter method. If done anyway, the request will fail.

Additional Parameters

$.xgafv (query-string) - V1 error format.

access_token (query-string) - OAuth access token.

alt (query-string) - Data format for response.

callback (query-string) - JSONP

fields (query-string) - Selector specifying which fields to include in a partial response.

key (query-string) - API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token.

oauth_token (query-string) - OAuth 2.0 token for the current user.

prettyPrint (query-boolean) - Returns response with indentations and line breaks.

quotaUser (query-string) - Available to use for quota purposes for server-side applications. Can be any arbitrary string assigned to a user, but should not exceed 40 characters.

uploadType (query-string) - Legacy upload protocol for media (e.g. "media", "multipart").

upload_protocol (query-string) - Upload protocol for media (e.g. "raw", "multipart").

pub fn add_scope<T, S>(
self,
scope: T
) -> ProjectLocationDatasetHl7V2StoreMessageListCall<'a, C> where
T: Into<Option<S>>,
S: AsRef<str>,

Identifies the authorization scope for the method you are building.

Use this method to actively specify which scope should be used, instead the default Scope variant Scope::CloudPlatform.

The scope will be added to a set of scopes. This is important as one can maintain access tokens for more than one scope. If None is specified, then all scopes will be removed and no default scope will be used either. In that case, you have to specify your API-key using the key parameter (see the param() function for details).

Usually there is more than one suitable scope to authorize an operation, some of which may encompass more rights than others. For example, for listing resources, a read-only scope will be sufficient, a read-write scope will do as well.